+++ frontend/src/domains/restaurantExclude/components/RestaurantItem.stories.tsx
