+++ frontend/src/domains/restaurantExclude/components/RestaurantList.stories.tsx
