+++ frontend/src/domains/restaurantExclude/components/RestaurantList.tsx
